update file naming convention and references
