+++ src/app/RecordDetailsEdit.tsx
@@ -21,11 +21,11 @@ import { Pressable } from "react-native";
 import FontAwesome6 from "@expo/vector-icons/FontAwesome6";
 import FontAwesome from "@expo/vector-icons/FontAwesome";
 import { RecordDetails } from "../types/RecordDetails";
-import { getId } from "../database/auth";
+import { getId } from "../lib/supabase";
 import { StyleSheet } from "react-native";
 import AlcoholItem from "../components/listSections/RecordDetails/AlcoholItem";
 import CaffeineItem from "../components/listSections/RecordDetails/CaffeineItem";
-import { getSleepRecordFromReadRecord } from "../utils/SleepRecord";
+import { getSleepRecordFromReadRecord } from "../utils/sleepRecord";
 import { TimeOfDay } from "../types/TimeOfDay";
 import { QualityOfSleep } from "../types/QualityOfSleep";
 import AfterSleepSection from "../components/listSections/RecordDetails/AfterSleepSection";
