+++ src/lib/supabase.ts
@@ -12,3 +12,11 @@ export const supabase = createClient(supabaseUrl, supabaseAnonKey, {
     detectSessionInUrl: false,
   },
 })
+
+export const getId = async (): Promise<string> => {
+  const { data: { user } } = await supabase.auth.getUser()
+  if (!user || !user.id) {
+    throw new Error("uuid is undefined");
+  }
+  return user.id;
+}
